+++ inc/toolbox.class.php
@@ -2499,25 +2499,6 @@ class Toolbox {
    }
 
    /**
-    * Remove accentued characters and return lower case string
-    *
-    * @param string $string String to handle
-    *
-    * @return string
-    */
-   public static function removeHtmlSpecialChars($string) {
-      $string = htmlentities($string, ENT_NOQUOTES, 'utf-8');
-      $string = preg_replace(
-         '#&([A-za-z])(?:acute|cedil|caron|circ|grave|orn|ring|slash|th|tilde|uml);#',
-         '\1',
-         $string
-      );
-      $string = preg_replace('#&([A-za-z]{2})(?:lig);#', '\1', $string);
-      $string = preg_replace('#&[^;]+;#', '', $string);
-      return self::strtolower($string, 'UTF-8');
-   }
-
-   /**
     * Slugify
     *
     * @param string $string String to slugify